Added Blockage interface
