+++ Chess/Chess.csproj
@@ -43,14 +43,13 @@
     <PlatformTarget>AnyCPU</PlatformTarget>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.ObjectPool" Version="5.0.0" />
+    <PackageReference Include="Microsoft.Extensions.ObjectPool" Version="5.0.1" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
     <PackageReference Include="System.Memory" Version="4.5.4" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Protocol\" />
     <Folder Include="Validation\" />
-    <Folder Include="Properties\" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\LICENSE">
